Move `WorkflowEnum` model to `warp_server_client`.
