+++ crates/warp_server_client/src/cloud_object/mod.rs
@@ -27,6 +27,7 @@ use crate::ids::{FolderId, ServerId, SyncId};
 mod creation;
 mod generic_cloud_object;
 mod generic_string_model;
+pub mod models;
 mod server_object;
 mod update;
